+++ empymod/utils.py
@@ -60,7 +60,7 @@ __all__ = ['EMArray', 'check_time_only', 'check_time', 'check_model',
            'check_bipole', 'check_ab', 'check_solution', 'get_abs',
            'get_geo_fact', 'get_azm_dip', 'get_off_ang', 'get_layer_nr',
            'printstartfinish', 'conv_warning', 'set_minimum', 'get_minimum',
-           'spline_backwards_hankel', 'Versions']
+           'spline_backwards_hankel', 'Report']
 
 # 0. General settings
 
@@ -1976,8 +1976,8 @@ def spline_backwards_hankel(ht, htarg, opt):
     return htarg, opt
 
 
-# 6. Versions
-class Versions(scooby.Report):
+# 6. Report
+class Report(scooby.Report):
     r"""Print date, time, and version information.
 
     Use scooby to print date, time, and package version information in any
@@ -2014,10 +2014,10 @@ class Versions(scooby.Report):
     --------
     >>> import pytest
     >>> import dateutil
-    >>> from emg3d import Versions
-    >>> Versions()                            # Default values
-    >>> Versions(pytest)                      # Provide additional package
-    >>> Versions([pytest, dateutil], ncol=5)  # Set nr of columns
+    >>> from emg3d import Report
+    >>> Report()                            # Default values
+    >>> Report(pytest)                      # Provide additional package
+    >>> Report([pytest, dateutil], ncol=5)  # Set nr of columns
 
     """